API con funcionalidades básicas de registro de usuarios y coordenadas finalizada.

+++ app/Http/Controllers/PointController.php
@@ -4,11 +4,21 @@ namespace App\Http\Controllers;
 
 use App\Models\Point;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-class PointController extends Controller
+class PointController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
+
     /**
-     * Display a listing of the resource.
+     * Muestra todos los puntos almacenados en la base de datos.
      */
     public function index()
     {
@@ -16,22 +26,22 @@ class PointController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Almacena un nuevo punto asociado a un usuario.
      */
     public function store(Request $request)
     {
         $fields = $request->validate([
-            'longitude' => 'required|numeric|min:-180|max:180',
             'latitude' => 'required|numeric|min:-90|max:90',
+            'longitude' => 'required|numeric|min:-180|max:180',
         ]);
 
-        $point = Point::create($fields);
+        $point = $request->user()->points()->create($fields);
 
-        return $point;
+        return ['point' => $point, 'user' => $point->user];
     }
 
     /**
-     * Display the specified resource.
+     * Muestra los datos de un punto específico.
      */
     public function show(Point $point)
     {
@@ -39,10 +49,15 @@ class PointController extends Controller
     }
 
     /**
-     * Update the specified resource in storage.
+     * Actualiza los datos de un punto específico.
+     * Aplica políticas de acceso para que el punto sólo pueda ser
+     * modificado por el usuario que lo creó.
      */
     public function update(Request $request, Point $point)
     {
+        // Aplica la política de acceso.
+        Gate::authorize('modify', $point);
+
         $fields = $request->validate([
             'longitude' => 'required|numeric|min:-180|max:180',
             'latitude' => 'required|numeric|min:-90|max:90',
@@ -54,12 +69,17 @@ class PointController extends Controller
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Elimina un punto específico.
+     * Aplica políticas de acceso para que el punto sólo pueda ser
+     * eliminado por el usuario que lo creó.
      */
     public function destroy(Point $point)
     {
+        // Aplica la política de acceso.
+        Gate::authorize('modify', $point);
+
         $point->delete();
 
-        return ['message' => 'El punto ha sido borrado.'];
+        return ['message' => 'El punto ha sido eliminado.'];
     }
 }
